refactor stamps flow to use built-in stack navigation
